Backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.32.3

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2

# PDF + NLP
PyPDF2==3.0.1
pdfplumber==0.11.4
sentence-transformers==3.0.1
numpy==1.26.4
